update: Unity
I just found out that there are two packages by which Unity can be installed: `ubuntu-unity-desktop` and `unity`. What I have done in this PR is update Unity to install the `unity` package and add other apps to the extra packages flag that better fit Unity's UI, like Rudra did for the 22.04 release of Ubuntu Unity. This makes it so that Unity isn't dependent on GNOME apps to keep it installed, and it makes the UI just look naturally nicer