feat!: remove packages
This is something I tested in my own fork for the Unity package until something more permanent is proposed. When installing `unity`, Ubuntu will install GNOME apps by default because that is something Ubuntu does. This addition will remove GNOME apps when Unity is chosen as the desktop environment, so the user doesn't have to go through the trouble of removing the package themselves.